+++ docs/components/Landing/Features/index.tsx
@@ -1,9 +1,12 @@
-import kitchn, { Container, Text } from "kitchn";
+import { AvatarGroup, Container, Image, Text } from "kitchn";
+import { useData } from "nextra/data";
 
 const Features = () => {
+  const { repo, contributors } = useData();
+
   return (
     <Container id={"features"} py={"extraLarge"} align={"center"} section>
-      <Text transform={"uppercase"} weight={"bold"} align={"center"} h3>
+      <Text transform={"uppercase"} weight={"bold"} align={"center"} h2>
         {"All features"}{" "}
         <Text transform={"uppercase"} weight={"bold"} accent={"primary"} span>
           {"Kitchn contains\r"}
@@ -22,8 +25,12 @@ const Features = () => {
       </Text>
       <Container mt={"large"} w={"100%"} gap={"normal"}>
         <Container direction={["column", "column", "row"]} gap={"normal"}>
-          <Container flex={1} gap={"normal"}>
-            <FeatureContainer flex={1} minH={200} p={"medium"}>
+          <Container
+            flex={1}
+            gap={"normal"}
+            direction={["column", "row", "column"]}
+          >
+            <Container br={"square"} bw={1} minH={200} p={"medium"}>
               <Text size={"large"} weight={"bold"}>
                 {"Accessible\r"}
               </Text>
@@ -32,18 +39,25 @@ const Features = () => {
                   "Kitchn strictly follows WAI-ARIA standards for all components.\r"
                 }
               </Text>
-            </FeatureContainer>
-            <FeatureContainer flex={1} minH={100} p={"medium"}>
+              <Image
+                mt={"small"}
+                src={"/static/images/features/accessible.svg"}
+                alt={"WAI-ARIA Accessible"}
+                w={"100%"}
+                h={"auto"}
+              />
+            </Container>
+            <Container br={"square"} bw={1} flex={1} minH={100} p={"medium"}>
               <Text size={"large"} weight={"bold"}>
                 {"Responsive support\r"}
               </Text>
               <Text mt={"small"} color={"lighter"}>
                 {"Build simple web apps with our simple responsive system.\r"}
               </Text>
-            </FeatureContainer>
+            </Container>
           </Container>
           <Container flex={1} gap={"normal"}>
-            <FeatureContainer flex={1} minH={100} p={"medium"}>
+            <Container br={"square"} bw={1} flex={1} minH={100} p={"medium"}>
               <Text size={"large"} weight={"bold"}>
                 {"Themeable\r"}
               </Text>
@@ -52,8 +66,8 @@ const Features = () => {
                   "Customize any part of our components to match your design needs.\r"
                 }
               </Text>
-            </FeatureContainer>
-            <FeatureContainer flex={1} minH={200} p={"medium"}>
+            </Container>
+            <Container br={"square"} bw={1} flex={1} minH={200} p={"medium"}>
               <Text size={"large"} weight={"bold"}>
                 {"All platform support\r"}
               </Text>
@@ -61,10 +75,14 @@ const Features = () => {
                 {"Whether you use React, React Native or a framework react,\r"}
                 {"Kitchn is compatible with all platforms.\r"}
               </Text>
-            </FeatureContainer>
+            </Container>
           </Container>
-          <Container flex={1} gap={"normal"}>
-            <FeatureContainer flex={1} minH={200} p={"medium"}>
+          <Container
+            flex={1}
+            gap={"normal"}
+            direction={["column", "row", "column"]}
+          >
+            <Container br={"square"} bw={1} minH={200} p={"medium"}>
               <Text size={"large"} weight={"bold"}>
                 {"Active Community\r"}
               </Text>
@@ -74,18 +92,56 @@ const Features = () => {
                 }
                 {"you need, and using it for our own products.\r"}
               </Text>
-            </FeatureContainer>
-            <FeatureContainer flex={1} minH={100} p={"medium"}>
+              <Container mt={"small"} br={"square"} h={"100%"}>
+                <Text color={"lighter"} weight={"medium"} size={"medium"} span>
+                  <Text color={"inherit"} weight={"bold"} size={"large"} span>
+                    {contributors.length}
+                  </Text>
+                  {" contributors"}
+                </Text>
+
+                <Text color={"lighter"} weight={"medium"} size={"medium"} span>
+                  <Text color={"inherit"} weight={"bold"} size={"large"} span>
+                    {repo.stargazers_count}
+                  </Text>
+                  {" stars"}
+                </Text>
+
+                <Text color={"lighter"} weight={"medium"} size={"medium"} span>
+                  <Text color={"inherit"} weight={"bold"} size={"large"} span>
+                    {new Date(repo.updated_at).toLocaleDateString()}
+                  </Text>
+                  {" last update"}
+                </Text>
+              </Container>
+              <AvatarGroup
+                mt={"small"}
+                members={contributors
+                  .filter((contributor) => !contributor.login.includes("[bot]"))
+                  .map((contributor) => ({
+                    src: contributor.avatar_url,
+                    alt: contributor.login,
+                  }))}
+              />
+            </Container>
+            <Container br={"square"} bw={1} minH={100} p={"medium"}>
               <Text size={"large"} weight={"bold"}>
                 {"Reusable hooks\r"}
               </Text>
               <Text mt={"small"} color={"lighter"}>
                 {"Lots of useful hooks to use in your applications.\r"}
               </Text>
-            </FeatureContainer>
+              <Image
+                mt={"small"}
+                src={"/static/images/features/reusable-hooks.svg"}
+                alt={"Reusable Hooks"}
+                w={"100%"}
+                h={"auto"}
+              />
+            </Container>
           </Container>
         </Container>
-        <FeatureContainer p={"medium"}>
+        <Container br={"square"} bw={1} p={"medium"}>
           <Text size={"large"} weight={"bold"}>
             {"Developer experience\r"}
           </Text>
@@ -95,9 +151,9 @@ const Features = () => {
             }
             {"or website.\r"}
           </Text>
-        </FeatureContainer>
+        </Container>
         <Container direction={["column", "row", "row"]} gap={"normal"}>
-          <FeatureContainer flex={1} p={"medium"}>
+          <Container br={"square"} bw={1} flex={1} p={"medium"}>
             <Text size={"large"} weight={"bold"}>
               {"Light & Dark support\r"}
             </Text>
@@ -106,8 +162,8 @@ const Features = () => {
                 "Kitchn strictly follows WAI-ARIA standards for all components.\r"
               }
             </Text>
-          </FeatureContainer>
-          <FeatureContainer flex={1} p={"medium"}>
+          </Container>
+          <Container br={"square"} bw={1} flex={1} p={"medium"}>
             <Text size={"large"} weight={"bold"}>
               {"Stylable\r"}
             </Text>
@@ -117,20 +173,18 @@ const Features = () => {
               }
               {"already implemented.\r"}
             </Text>
-          </FeatureContainer>
+            <Image
+              mt={"small"}
+              src={"/static/images/features/stylable.svg"}
+              alt={"Stylable"}
+              w={"100%"}
+              h={"auto"}
+            />
+          </Container>
         </Container>
       </Container>
     </Container>
   );
 };
-
-const FeatureContainer = kitchn(Container)`
-  border-radius: ${({ theme }) => theme.radius.square};
-  background-color: ${({ theme }) => theme.colors.layout.darker};
-
-  @media (min-width: ${({ theme }) => theme.breakpoint.mobile}) {
-    
-  }
-`;
 
 export default Features;