+++ docs/components/Landing/Showcase/List/Card/index.tsx
@@ -1,53 +1,42 @@
-import kitchn, { Image, defaultThemes } from "kitchn";
-import { Showcase } from "types/showcase";
+import kitchn from "kitchn";
+import { Company, SortedProject } from "types/showcase";
 
 export type ShowcaseCardProps = {
-  name: Showcase["name"];
-  thumbnail: Showcase["thumbnail"];
+  company: Company | SortedProject["owner"];
 };
 
 const ShowcaseCard: React.FC<ShowcaseCardProps> = ({
-  name,
-  thumbnail,
-  ...props
+  company,
 }: ShowcaseCardProps) => {
+  const { logo: Logo } = company;
+
   return (
-    <Container title={name} {...props}>
-      <ThumbnailImage
-        src={`/static/images/showcases/thumbnails/${thumbnail}`}
-        alt={name}
-        height={50}
-        width={150}
-      />
+    <Container>
+      <Logo />
     </Container>
   );
 };
 
-const ThumbnailImage = kitchn(Image)`
-  opacity: 0.3;
-  width: 100%;
-  max-width: 60%;
-  height: 100%;
-  max-height: 24px;
-  transition: all 1s;
-  filter: ${({ theme }) =>
-    theme.name === defaultThemes.light.name &&
-    "grayscale(0) brightness(0) invert(0)"};
-`;
+const Container = kitchn.span`
+  position: relative;
+  height: 2rem;
+  margin: 0 1rem;
+  bottom: 0;
+  opacity: 0.8;
+  filter: grayscale(1);
+  transition: opacity 125ms ease-in-out;
 
-const Container = kitchn.div`
-  width: 160px;
-  height: 64px;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  transition: all 1s;
+  .dark & {
+    filter: brightness(0) invert(1);
+  }
+
+  svg {
+    height: 100%;
+    max-width: 128px;
+  }
 
   &:hover {
-    transform: scale(1.05);
-    ${ThumbnailImage} {
-      opacity: 1;
-    }
+    opacity: 1;
   }
 `;