+++ packages/kitchn/src/components/Image/index.tsx
@@ -1,12 +1,25 @@
 import React from "react";
 import styled from "styled-components";
 
-import { withDecorator } from "../../hoc/withDecorator";
+import { AreaProps, withDecorator } from "../../hoc/withDecorator";
 import { KitchnComponent } from "../../types";
 
 type Props = {
+  width?: AreaProps["width"];
+  w?: Props["width"];
+  height?: AreaProps["height"];
+  h?: Props["height"];
   src: string;
   alt: string;
+  objectFit?:
+    | "cover"
+    | "contain"
+    | "fill"
+    | "scale-down"
+    | "inherit"
+    | "initial"
+    | "unset"
+    | "none";
 };
 
 export type ImageProps = KitchnComponent<
@@ -16,7 +29,9 @@ export type ImageProps = KitchnComponent<
 
 const ImageComponent = styled(({ src, alt, ...props }: ImageProps) => {
   return <img src={src} alt={alt} draggable={false} {...props} />;
-})<ImageProps>``;
+})<ImageProps>`
+  ${({ objectFit }) => objectFit && `object-fit: ${objectFit};`}
+`;
 
 ImageComponent.displayName = "KitchnImage";
 export const Image = withDecorator(ImageComponent);